move util into utils file
